pyproject.toml
[project]
name = "sitcom-ml-project"
version = "0.1.0"
description = "Mood-Moment Recommender — semantic TV episode discovery"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "python-dotenv>=1.0",
    "langgraph>=0.2.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4"]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"
